Move the dashboard over into personalization folder
