+++ config/webpack.config.js
@@ -24,6 +24,7 @@ const entryFiles = {
   '5495-edu-benefits': './src/js/edu-benefits/5495/edu-benefits-entry.jsx',
   '526EZ-claims-increase': './src/js/disability-benefits/526EZ/form-entry.jsx',
   'claims-status': './src/js/claims-status/claims-status-entry.jsx',
+  'dashboard-beta': './src/js/personalization/dashboard-beta/dashboard-entry.jsx',
   'discharge-upgrade-instructions': './src/js/discharge-wizard/discharge-wizard-entry.jsx',
   'health-records': './src/js/health-records/health-records-entry.jsx',
   'id-card-beta': './src/js/id-card-beta/id-card-beta-entry.jsx',
@@ -37,7 +38,6 @@ const entryFiles = {
   'account-beta': './src/js/personalization/account-beta/account-entry.jsx',
   auth: './src/js/auth/auth-entry.jsx',
   burials: './src/js/burials/burials-entry.jsx',
-  dashboard: './src/js/dashboard/dashboard-entry.jsx',
   facilities: './src/js/facility-locator/facility-locator-entry.jsx',
   gi: './src/js/gi/gi-entry.jsx',
   hca: './src/js/hca/hca-entry.jsx',
