+++ src/js/personalization/dashboard-beta/actions/index.js
@@ -1,6 +1,6 @@
-import { removeFormApi } from '../../common/schemaform/save-in-progress/api';
-import { apiRequest } from '../../common/helpers/api';
-import { getUserData } from '../../common/helpers/login-helpers';
+import { removeFormApi } from '../../../common/schemaform/save-in-progress/api';
+import { apiRequest } from '../../../common/helpers/api';
+import { getUserData } from '../../../common/helpers/login-helpers';
 
 export const UPDATE_PROFILE_FIELDS = 'UPDATE_PROFILE_FIELDS';
 export const PROFILE_LOADING_FINISHED = 'PROFILE_LOADING_FINISHED';
