+++ src/js/personalization/dashboard-beta/components/AuthApplicationSection.jsx
