+++ src/js/personalization/dashboard-beta/components/FormList.jsx
@@ -1,8 +1,8 @@
 import PropTypes from 'prop-types';
 import React from 'react';
 
-import ProgressButton from '../../common/components/form-elements/ProgressButton';
-import Modal from '../../common/components/Modal';
+import ProgressButton from '../../../common/components/form-elements/ProgressButton';
+import Modal from '../../../common/components/Modal';
 
 import FormItem from './FormItem';
 import { isSIPEnabledForm } from '../helpers';
