+++ src/js/personalization/dashboard-beta/containers/DashboardApp.jsx
@@ -2,16 +2,16 @@ import React from 'react';
 import { connect } from 'react-redux';
 import moment from 'moment';
 
-import { getVerifyUrl } from '../../common/helpers/login-helpers.js';
-import { updateVerifyUrl } from '../../login/actions';
+import { getVerifyUrl } from '../../../common/helpers/login-helpers.js';
+import { updateVerifyUrl } from '../../../login/actions';
 import { removeSavedForm } from '../actions';
 
 import FormList from '../components/FormList';
 import MessagingWidget from './MessagingWidget';
 import ClaimsAppealsWidget from './ClaimsAppealsWidget';
 
-import RequiredLoginView from '../../common/components/RequiredLoginView';
-import DowntimeNotification, { services } from '../../common/containers/DowntimeNotification';
+import RequiredLoginView from '../../../common/components/RequiredLoginView';
+import DowntimeNotification, { services } from '../../../common/containers/DowntimeNotification';
 
 moment.updateLocale('en', {
   meridiem: (hour) => {
