+++ src/js/personalization/dashboard-beta/dashboard-entry.jsx
@@ -1,7 +1,7 @@
 import 'core-js';
-import '../common';
-import '../../sass/user-profile.scss';
-import '../../sass/dashboard.scss';
+import '../../common';
+import '../../../sass/user-profile.scss';
+import '../../../sass/dashboard.scss';
 
 import React from 'react';
 import ReactDOM from 'react-dom';
@@ -9,9 +9,9 @@ import ReactDOM from 'react-dom';
 import { Router, browserHistory } from 'react-router';
 import { Provider } from 'react-redux';
 
-import initReact from '../common/init-react';
+import initReact from '../../common/init-react';
 import routes from './routes';
-import initCommon from '../common/init-common';
+import initCommon from '../../common/init-common';
 import reducer from './reducers';
 
 const commonStore = initCommon(reducer);
