+++ src/js/personalization/dashboard-beta/reducers/index.js
